Address review comments: pass wallet mode in query parameters instead of sessionStorage
